% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accepted_model}
\alias{accepted_model}
\title{copy the previous SSC accepted model as the new "base" model}
\usage{
accepted_model(base_year, base_model, year, folder = NULL)
}
\arguments{
\item{base_year}{year of the base model}

\item{base_model}{name of the base model (folder name)}

\item{year}{current year}

\item{folder}{if the base model is in a different location than the afscdata structure is looking for}
}
\value{
a 'base' model folder with all of the prior years inputs/outputs
}
\description{
copy of all base model files, adds a "README.md" file that identifies the model used (or appends the current README), predicated on using the 'afscdata::setup_folders()' function, though has an option for other folder structures. Note, this function copies everything over (e.g., retrospectives) and may take a few minutes to complete
}
\examples{
\dontrun{
accepted_model(base_year = 2020, base_model = "m18.2b", year = 2021)
}
}
