% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afscdata-package.R
\docType{package}
\name{afscdata-package}
\alias{afscdata}
\alias{afscdata-package}
\title{afscdata: An R Package for Extraction of AFSC Fishery Data}
\description{
This package provides a folder structure and SQL queries for accessing the AKFIN and AFSC databases in support of fishery stock assessments. Additionally, archived data are held/provided to maintain consistency across assessments/assessment authors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://afsc-assessments.github.io/afscdata}
}

}
\author{
\strong{Maintainer}: Benjamin Williams \email{ben.williams@noaa.gov}

Authors:
\itemize{
  \item Benjamin Williams \email{ben.williams@noaa.gov}
}

Other contributors:
\itemize{
  \item Jane Sullivan \email{jane.sullivan@noaa.gov} [contributor]
  \item Steve Barbeaux [contributor]
  \item Pete Hulson [contributor]
  \item Jim Ianelli [contributor]
  \item Matt Callahan [contributor]
  \item Meaghan Bryan [contributor]
  \item Kalei Shotwell [contributor]
  \item Margaret Siple [contributor]
}

}
\keyword{internal}
