% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atka_mackerel.R
\name{bsai_amak}
\alias{bsai_amak}
\title{raw data query for the BSAI Atka mackerel assessment}
\usage{
bsai_amak(year, off_yr = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for the BSAI Atka mackerel assessment
}
\examples{
\dontrun{
bsai_amak(year = 2023, off_yr = TRUE)
}
}
