% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsai_amak_catch_1977_1990}
\alias{bsai_amak_catch_1977_1990}
\title{BSAI Atka mackerel historical catch}
\format{
A data frame with 14 observations and 2 variables:
\describe{
\item{year}{1977-1990}
\item{catch}{catch in tons}
}
}
\usage{
bsai_amak_catch_1977_1990
}
\description{
A dataset containing historical catch data by year for 1977-1990, pulled from
2022 stock assessment data file.
}
\keyword{datasets}
