% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsai_flats.R
\name{bsai_fhs}
\alias{bsai_fhs}
\title{raw data query for BSAI flathead sole}
\usage{
bsai_fhs(year, off_yr = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for BSAI flathead sole
}
\examples{
\dontrun{
bsai_fhs(year = 2022, off_yr = FALSE)
}
}
