% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsai_fhs_catch_1964_1994}
\alias{bsai_fhs_catch_1964_1994}
\title{BSAI flathead sole historical catch data 1964-1994}
\format{
A data frame with 30 observations and 2 variables:
\describe{
\item{Year}{year of catch}
\item{Catch}{weight of catch in 1,000 t}
}
}
\usage{
bsai_fhs_catch_1964_1994
}
\description{
A dataset containing trawl gear catch by year for 1964-1994
}
\keyword{datasets}
