% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_6.R
\name{bsai_octopus}
\alias{bsai_octopus}
\title{raw data query for BSAI octopus}
\usage{
bsai_octopus(year)
}
\arguments{
\item{year}{assessment year}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for BSAI octopus
}
\examples{
\dontrun{
bsai_octopus(year = 2023)
}
}
