% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsai_rockfish.R
\name{bsai_orox}
\alias{bsai_orox}
\title{raw data query for BSAI other rockfish}
\usage{
bsai_orox(year, off_yr = FALSE, catch_report = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}

\item{catch_report}{is this a catch report year, default FALSE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for BSAI other rockfish
}
\examples{
\dontrun{
bsai_orox(year = 2022, off_yr = FALSE)
}
}
