% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_filters}
\alias{collapse_filters}
\title{utility function for sql queries}
\usage{
collapse_filters(x)
}
\arguments{
\item{x}{variable to add quotes to}
}
\description{
adds correct quotes for sql queries, nested within \code{sql_filter}
}
