% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect}
\alias{connect}
\title{utility function to connect to server}
\usage{
connect(db = "akfin")
}
\arguments{
\item{db}{the database schema ("akfin" or "afsc")}
}
\description{
utility function to connect to server
}
