% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disconnect}
\alias{disconnect}
\title{utility function to disconnect from server}
\usage{
disconnect(db)
}
\arguments{
\item{db}{the database schema (e.g., akfin or afsc)}
}
\description{
utility function to disconnect from server
}
