% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmp_key}
\alias{fmp_key}
\title{Sablefish fmp area key by location}
\format{
A data frame with 923 observations and 2 variables:
\describe{
\item{loc}{long_lat}
\item{FMP_SUBAREA}{defined area for location}
}
}
\usage{
fmp_key
}
\description{
A dataset containing area definitions
}
\keyword{datasets}
