% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_check_bio.R
\name{gap_check_bio}
\alias{gap_check_bio}
\title{compare GAP updated survey biomass to (now) retired design-based biomass estimates}
\usage{
gap_check_bio(year, species, area, type)
}
\arguments{
\item{year}{current year}

\item{species}{afsc species codes e.g., 30420}

\item{area}{options = ai, goa, ebs, bss, nbs}

\item{type}{= region, subarea, area, stat_area, stratum, inpfc, inpfc_depth, depth, reg_area_depth}
}
\value{
a list with orig values, gap values, and a basic report
}
\description{
compare GAP updated survey biomass to (now) retired design-based biomass estimates
}
\examples{
\dontrun{
out <- gap_check_bio(year = 2024, species = 30420, type = 'total', area = 'AI' )
out$report
}
}
