% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atka_mackerel.R
\name{goa_amak}
\alias{goa_amak}
\title{raw data query for the GOA Atka mackerel assessment}
\usage{
goa_amak(year)
}
\arguments{
\item{year}{assessment year}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for the GOA Atka mackerel assessment
}
\examples{
\dontrun{
}
}
