% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_rockfish.R
\name{goa_dusk}
\alias{goa_dusk}
\title{raw data query for GOA dusky rockfish}
\usage{
goa_dusk(year, off_yr = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for GOA dusky rockfish
}
\examples{
\dontrun{
goa_dusk(year = 2022, off_yr = FALSE)
}
}
