% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_fhs_catch_1978_1990}
\alias{goa_fhs_catch_1978_1990}
\title{GOA flathead sole historical catch}
\format{
A data frame with 13 observations and 2 variables:
\describe{
\item{year}{1978-1990}
\item{catch}{catch in tons}
}
}
\usage{
goa_fhs_catch_1978_1990
}
\description{
A dataset containing historical catch data by year for 1978-1990
}
\keyword{datasets}
