% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pcod_catch_1977_1990}
\alias{goa_pcod_catch_1977_1990}
\title{GOA Pacific cod historical larval index data}
\format{
A data frame with 92 observations and 5 variables:
\describe{
\item{year}{1977-2022}
\item{seas}{season?}
\item{index}{index id}
\item{obs}{observation}
\item{se_log}{log standard error of obs}
}
}
\usage{
goa_pcod_catch_1977_1990
}
\description{
A dataset containing larval index collections year for 1977-2022
}
\keyword{datasets}
