% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_rockfish.R
\name{goa_pop}
\alias{goa_pop}
\title{raw data query for GOA POP}
\usage{
goa_pop(year, off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for GOA POP
}
\examples{
\dontrun{
goa_pop(year = 2022, off_yr = FALSE)
}
}
