% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pop_catch_1960_1990}
\alias{goa_pop_catch_1960_1990}
\title{GOA POP historical catch data}
\format{
A data frame with 30 observations and 2 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in 1,000 t}
}
}
\usage{
goa_pop_catch_1960_1990
}
\description{
A dataset containing trawl gear catch by year for 1961-1990
}
\keyword{datasets}
