% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pop_fixed_fish_length_comp}
\alias{goa_pop_fixed_fish_length_comp}
\title{GOA POP historical fishery length comp data}
\format{
A data frame with 765 observations and 4 variables:
\describe{
\item{Length}{length in cm}
\item{value}{number of lengths collected}
\item{year}{1963-1977}
\item{hauls_sampled}{number of hauls sampled}
}
}
\usage{
goa_pop_fixed_fish_length_comp
}
\description{
A dataset containing length comp collection data by year for 1963-1977
}
\keyword{datasets}
