% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_rebs_catch_1977_2004}
\alias{goa_rebs_catch_1977_2004}
\title{GOA rougheye/blackspotted rockfish historical catch data}
\format{
A data frame with 28 observations and 2 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in t}
}
}
\usage{
goa_rebs_catch_1977_2004
}
\description{
A dataset containing trawl and longline gear catch by year for 1977-2004
}
\details{
The catches from 1977-1992 were from Soh (1998), which reconstructs the catch
history using an information weighting factor (λ) to combine catch histories
from both survey and fishery information. The catches from 1993-2004 were
constructed  used observer catch data from the FMA Observer Program (Clausen
et al. 2004, Appendix A). Observed catches were available from the FMA
database by area, gear, and species for hauls sampled by observers. This
information was used to calculate proportions of RE/BS catch by gear type.
These proportions were then applied to the combined shortraker/rougheye catch
from the NMFS Alaska Regional Office to yield estimates of total catch for
RE/BS rockfish.

Clausen, D. M., D.H. Hanselman, J.T. Fujioka, and J. Heifetz.  2004. Gulf of
Alaska shortraker/rougheye and other slope rockfish. In Stock assessment and
fishery evaluation report for the groundfish resources of the Gulf of Alaska,
p. 413 – 463. North Pacific Fishery Management Council, 605 W 4th Ave, Suite
306, Anchorage AK 99501. https://apps-afsc.fisheries.noaa.gov/refm/docs/2004/GOAsloperf.pdf

Soh, Sung Kwon. 1998. The use of harvest refugia in the management of
shortraker and rougheye rockfish (Sebastes borealis/Sebastes aleutianus) in
the Gulf of Alaska. Ph.D. Thesis – University of Washington. 194 pp.
}
\keyword{datasets}
