% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_flats.R
\name{goa_shal_flats}
\alias{goa_shal_flats}
\title{raw data query for GOA shallow-water flatfish}
\usage{
goa_shal_flats(year, off_yr = FALSE, catch_report = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}

\item{catch_report}{is this a catch report year, default FALSE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for GOA shallow-water flatfish
}
\examples{
\dontrun{
goa_shal_flats(year = 2022, off_yr = FALSE)
}
}
