% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_rockfish.R
\name{goa_thornyhead}
\alias{goa_thornyhead}
\title{raw data query for GOA thorynhead rockfish}
\usage{
goa_thornyhead(year, off_yr = FALSE, catch_report = FALSE)
}
\arguments{
\item{year}{assessment year}

\item{off_yr}{if this is an off-year assessment change to TRUE}

\item{catch_report}{is this a catch report year, default FALSE}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for GOA thorynhead rockfish
}
\examples{
\dontrun{
goa_thornyhead(year = 2022, off_yr = FALSE)
}
}
