% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_biomass}
\alias{q_bts_biomass}
\title{query bottom trawl survey biomass}
\usage{
q_bts_biomass(
  year,
  species,
  area,
  type = 1,
  db,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit afsc species code(s) e.g., 79210 or c(79210, 90210)}

\item{area}{options are ebs (the ebs+nw), bsslope, nbs, ai, goa, old_bs (was called "standard")}

\item{type}{area numeric values - can only use a single switch, default = 1 (region)
1 REGION - all
2 REGULATORY AREA - goa/ai
3 STRATUM - goa/ai/nbs
4 INPFC - goa
5 INPFC BY DEPTH -goa
6 DEPTH - all(except nbs)
7 SUBAREA - ai
8 SUBAREA BY DEPTH - ai}

\item{db}{the database to query (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\value{
saves bts biomass data as data/raw/area_(type)_bts_biomass_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
probably need to beef up the documentation on the "by" switch
}
\details{
six areas are available to query from
ebs = bering sea + northwest 1987-present (includes nw stations) - recommended
bsslope = bering sea slope
nbs = northern bering sea
ai = aleutian islands
goa = gulf of alaska
old_bs = bering sea standard 1982-present (minus ~20 stations in nw) - not recommended

for the goa and ai there is a "type" switch that queries by depth, stratum, area, total, or uses the inpfc table or inpfc by depth table only one of these can be used at a time.
}
\examples{
\dontrun{
db <- afscdata::connect("akfin")
q_bts_biomass(year=2022, species=21921, area = "goa", type = 1, db = db)
} 
}
