% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_gap_agecomp}
\alias{q_bts_gap_agecomp}
\title{query GAP bottom trawl survey agecomps}
\usage{
q_bts_gap_agecomp(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit species code (e.g., 10110) - can place multiple in a vector c(10110, 10130)}

\item{area}{options: ebs, ai, goa, nbs, ebs_slope, ebs_nbs, default: goa}

\item{db}{data server to connect to (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: true)}
}
\value{
saves bts length data as data/raw/bts_gap_agecomp_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
query GAP bottom trawl survey agecomps
}
\examples{
\dontrun{
db <- afscdata::connect()
q_bts_gap_agecomp(year=2022, species=21921, area = "goa", db = db)
} 
}
