% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_gap_sizecomp}
\alias{q_bts_gap_sizecomp}
\title{query GAP bottom trawl survey sizecomps}
\usage{
q_bts_gap_sizecomp(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit afsc species code(s) e.g., 79210 or c(79210, 90210)}

\item{area}{options: ebs, ai, goa, nbs, ebs_slope, ebs_nbs, default: goa}

\item{db}{the database to query (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\value{
saves bts sizecomp data as data/raw/area_bts_gap_sizecomp_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
query GAP bottom trawl survey sizecomps
}
\examples{
\dontrun{
db <- afscdata::connect()
q_bts_gap_sizecomp(year=2022, species=21921, area = "goa", db = db)
} 
}
