% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_length}
\alias{q_bts_length}
\title{query bottom trawl survey length data from the AFSC server}
\usage{
q_bts_length(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit species code (e.g., 10110) - can place multiple in a vector c(10110, 10130)}

\item{area}{bs, hwc, wc, hg, ai, goa, or hbs}

\item{db}{data server to connect to (afsc)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: true)}
}
\description{
query bottom trawl survey length data from the AFSC server
}
