% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_sizecomp}
\alias{q_bts_sizecomp}
\title{query bottom trawl survey sizecomps}
\usage{
q_bts_sizecomp(
  year,
  species,
  area,
  type = "total",
  db,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit afsc species code(s) e.g., 79210 or c(79210, 90210)}

\item{area}{options are 'ai' or 'goa' - can only call a single area}

\item{type}{"stratum" or "total" - only available for goa/ai (default: "total") - can only use a single switch}

\item{db}{the database to query (afsc)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\value{
saves bts sizecomp data as data/raw/area_(type)_bts_sizecomp_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
currently only available for goa and ai
}
\details{
for the goa and ai there is a "type" switch that queries by stratum, or total - !! stratum should not be used at the moment !!. only one of these can be used at a time.
}
\examples{
\dontrun{
db <- afscdata::connect("afsc")
q_bts_sizecomp(year=2022, species=21921, area = "goa", type = "total", db = db)
} 
}
