% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_specimen}
\alias{q_bts_specimen}
\title{query bottom trawl survey specimen data from the AFSC server}
\usage{
q_bts_specimen(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{5 digit species code (e.g., 10110) - can place multiple in a vector c(10110, 10130)}

\item{area}{bs, ai, goa, or hwc, wc, hg, hbs - can do multiples c("bs","ai")}

\item{db}{data server to connect to (afsc)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: true)}
}
\value{
saves bts length data as data/raw/bts_length_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
query bottom trawl survey specimen data from the AFSC server
}
\examples{
\dontrun{
db <- afscdata::connect("afsc")
q_bts_length(year=2022, species=21921, area = "goa", db = db)
} 
}
