% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_catch}
\alias{q_catch}
\title{query fishery catch data}
\usage{
q_catch(
  year,
  species,
  area,
  db,
  add_fields = NULL,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{species group code e.g., "DUSK" or numeric agency values e.g. c("131", "132") - must be either all 4 digit or 3 digit codes}

\item{area}{fmp_area (GOA, BSAI) or fmp_subarea (BS, AI, WG, CG, WY, EY, SE) - also available (SEI, PWSI)}

\item{db}{data server to connect to (akfin)}

\item{add_fields}{add other columns to the database (must currently exist on server). "*" will return all table columns available}

\item{print_sql}{outputs the sql query instead of calling the data}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\description{
Note that this function will produce results that are confidential. They can be hidden from git/GitHub by adding fsh_catch_data.csv to your .gitignore file
}
