% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_catch_foreign}
\alias{q_catch_foreign}
\title{query foreign catch data from the AKFIN server}
\usage{
q_catch_foreign(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{common name (e.g., "sablefish" or "all flounders") - can call multiple species}

\item{area}{numeric area digit code - multiples is ok}

\item{db}{data server to connect to (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false) - save must be false}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: true)}
}
\value{
saves catch data as data/raw/for_catch_data.csv or outputs to the global environment, also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
something about the species names and area codes, since they are so different from everything else?
pulls data from AKFIN pre1991.foreign_blend table
}
\examples{
\dontrun{
db <- afscdata::connect()
q_catch_foreign(year=2022, species="sablefish", area = 54:57, db = db)
}
}
