% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_fish_obs.R
\name{q_fish_obs}
\alias{q_fish_obs}
\title{fishery observer data query}
\usage{
q_fish_obs(year, fishery = "fsh", norpac_species, area, db, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh, change if multiple fisheries}

\item{norpac_species}{norpac species code}

\item{area}{goa or bsai}

\item{db}{the database to query}

\item{save}{save the file in designated folder or output to global environment}
}
\description{
fishery observer data query
}
\examples{
q_fish_obs(2022, norpac_species = 303, area = 'GOA', db = akfin, save = TRUE)

}
