% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_fish_ticket}
\alias{q_fish_ticket}
\title{query fish ticket data}
\usage{
q_fish_ticket(
  year,
  area,
  species,
  db,
  add_fields = NULL,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to query through (and location to save results)}

\item{area}{fmp_area (GOA, BSAI) or fmp_subarea (BS, AI, WG, CG, WY, EY, SE) - also available (SEI, PWSI), can use all fmp_areas or all fmp_subareas, but don't mix the two}

\item{species}{species group code e.g., "DUSK"}

\item{db}{data server to connect to (akfin)}

\item{add_fields}{add other columns to the database (must currently exist on server). "*" will return all table columns available}

\item{print_sql}{outputs the sql query instead of calling the data - save must be false}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\description{
query fish ticket data
}
