% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_lls_rpn}
\alias{q_lls_rpn}
\title{query nfms longline survey relative population numbers or weights}
\usage{
q_lls_rpn(
  year,
  area = c("goa", "bs", "ai"),
  species,
  by = "fmpsubarea",
  use_historical = FALSE,
  db,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{area}{options are 'goa', 'bs', 'ai', or a combo. default=c('goa', 'bs', 'ai')}

\item{species}{5 digit afsc/race species code(s) e.g., 79210 or c(79210, 10110)}

\item{by}{'depth' (stratum-level), 'geoarea' (e.g. Spencer Gully, Kodiak
slope), 'councilarea' (e.g., West Yakutat, East Yakutat/Southeast),
'fmpsubarea' (e.g., Eastern Gulf of Alaska), or 'akwide' (only for
sablefish). default: 'fmpsubarea' - can only call a single area}

\item{use_historical}{T/F include historical Japanese survey data in the
results (default: false)}

\item{db}{the database to query (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false) - save must be false}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\description{
available source tables on akfin:
lls_area_stratum_rpn
lls_area_stratum_rpn_depred
lls_area_rpn_all_strata
lls_area_rpn_3_to_7
lls_area_rpn_3_to_7_depred
lls_council_sablefish_area_all_strata
lls_council_sablefish_area_3_to_7_depred
lls_fmp_subarea_all_strata
lls_fmp_subarea_3_to_7_depred
lls_fmp_all_strata* not available in this fcn
lls_ak_wide_3_to_7_depred
}
