% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_lls_sable_specimen}
\alias{q_lls_sable_specimen}
\title{query nmfs longine survey specimen data (age, length, weight, sex, maturity).
only available for sablefish!}
\usage{
q_lls_sable_specimen(
  year,
  area = c("goa", "bs", "ai"),
  use_historical = FALSE,
  db = akfin,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{area}{options are 'goa', 'bs', 'ai', or a combo. default=c('goa', 'bs', 'ai')}

\item{use_historical}{T/F include historical Japanese survey data in the
results (default: false)}

\item{db}{the database to query (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false) - save must be false}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\value{
saves lls sablefish specimen data as data/raw/lls_specimen_data.csv or outputs to
the global environment. also saves a copy of the SQL code used for the
query and stores it in the data/sql folder.
}
\description{
longline database documentation available on
\href{https://akfinbi.psmfc.org/analyticsRes/Documentation/Database_Background_Instructions_AKFIN_20210915.pdf}{akfin}
}
\details{
source table on akfin:
afsc.lls_age_view
}
\examples{
\dontrun{
# sablefish specimen data in 1996 (first year ages were collected in the goa domestic survey) in the gulf of alaska
db <- afscdata::connect("akfin")
q_lls_specimen(year = 2000, area = "goa", use_historical = FALSE, db = db, save = FALSE)
} 
}
