% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{q_nontarget}
\alias{q_nontarget}
\title{query non-target species catch estimate}
\usage{
q_nontarget(year, target, area, db, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{target}{targeted species: 'p' = pollock-mid, 'b' = pollock-bottom, x' = rex, 'h' = shallow flats, 'k' = rockfish, 'w' = arrowtooth, 'c' = pcod, 'i' = halibut}

\item{area}{fmp_area (GOA, BSAI) or fmp_subarea (BS, AI, WG, CG, WY, EY, SE) - also available (SEI, PWSI)}

\item{db}{data server to connect to (akfin)}

\item{save}{save the file in designated folder (default = T) or the global environment}
}
\value{
a csv of non-target species catch by trip target group, saved in the data/output folder
}
\description{
non-target catch estimates by weight (or numbers)
}
\examples{
\dontrun{
akfin = afscdaya::connect()
q_nontarget(year=2022, target="k", area="goa", db=akfin, save=FALSE)
disconnect(akfin)
}

}
