% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{q_psc}
\alias{q_psc}
\title{query prohibited species catch (psc) estimate}
\usage{
q_psc(year, trip_target, area, db, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{trip_target}{'p' = pollock-mid, 'b' = pollock-bottom, x' = rex, 'h' = shallow flats, 'k' = rockfish, 'w' = arrowtooth, 'c' = pcod, 'i' = halibut}

\item{area}{goa or bsai}

\item{db}{data server to connect to (akfin)}

\item{save}{save the file in designated folder (default = T) or the global environment}
}
\value{
a csv of prohibited species catch by trip target group, saved in the data/output folder
}
\description{
prohibited species catch (PSC) estimates reported in tons for halibut and herring, counts for salmon, crabs and other fish. Note that you can combine trip target codes c("k", "x") and regions - though results will be lumped together
}
\examples{
\dontrun{
akfin = connect()
q_psc(year=2022, trip_target="k", area="goa", db, save=FALSE)
disconnect(akfin)
}

}
