% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_specs}
\alias{q_specs}
\title{Query specs}
\usage{
q_specs(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{for specifying the correct folder}

\item{species}{species group code e.g., "DUSK", "PCOD"}

\item{area}{"GOA" or "BSAI"}

\item{db}{data server to connect to (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data - save must be false}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default TRUE)}
}
\description{
Query specs
}
\examples{
\dontrun{
db = connect()
q_specs(year=2024, species='PCOD', area='BSAI', db=db, save=F)
}
}
