% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_ts_biomass}
\alias{q_ts_biomass}
\title{query bottom trawl survey biomass}
\usage{
q_ts_biomass(
  year,
  area = "goa",
  afsc_species,
  shelf_slope = "shelf",
  depth = NULL,
  stratum = NULL,
  akfin,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa"}

\item{afsc_species}{afsc species code(s)}

\item{shelf_slope}{if the area is the Bering Sea the default is designated "shelf" alternate is "slope"}

\item{depth}{default FALSE, otherwise queries data by depth}

\item{stratum}{default FALSE, otherwise queries data by stratum}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}

\item{db}{the database to query (akfin)}
}
\description{
query bottom trawl survey biomass
}
