% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saa_pop_60}
\alias{saa_pop_60}
\title{GOA POP historical size at age parameters data}
\format{
A data frame with 1 observations and 5 variables:
\describe{
\item{linf}{von B parameter}
\item{k}{von B parameter}
\item{t0}{von B parameter}
\item{a}{a parameter}
\item{b}{b parameter}
}
}
\usage{
saa_pop_60
}
\description{
A dataset containing size at age parameters for POP <= 1960
}
\keyword{datasets}
