% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_waa}
\alias{sabl_waa}
\title{Sablefish fixed weights at age}
\format{
A data frame with 90 observations and 3 variables:
\describe{
\item{waa}{weight at age}
\item{sex}{male or female}
\item{timeblock}{current or old}
}
}
\usage{
sabl_waa
}
\description{
A dataset containing weights at age for two time blocks, by sex
}
\keyword{datasets}
