% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{Setup folder structure}
\usage{
setup_folders(year, tier = NULL)
}
\arguments{
\item{year}{assessment year}

\item{tier}{assessment tier to change the folders used - not currently implemented}
}
\value{

}
\description{
Creates a common folder structure for assessment data
}
\examples{
\dontrun{
setup(2022)
}
}
