% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharks.R
\name{sharks}
\alias{sharks}
\title{raw data query for GOA/BSAI sharks}
\usage{
sharks(year, area)
}
\arguments{
\item{year}{assessment year}

\item{area}{what region? eith goa or bsai}
}
\value{
a suite of raw data .csv files and a time stamp of when the query was done
}
\description{
raw data query for GOA/BSAI sharks
}
\examples{
\dontrun{
sharks(year = 2023, area = "goa")
}
}
