% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_switch.R
\name{species_switch}
\alias{species_switch}
\title{assign species codes}
\usage{
species_switch(species, area)
}
\arguments{
\item{species}{species group name e.g., "DUSK"}

\item{area}{area of interest ("GOA", "BS" or "AI")}
}
\value{

}
\description{
assign species codes
}
\examples{
\dontrun{
species_switch("DUSK", "GOA")
}
}
