% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_filter}
\alias{sql_filter}
\title{utility function to filter sql files}
\usage{
sql_filter(sql_precode = "IN", x, sql_code, flag = "-- insert species")
}
\arguments{
\item{sql_precode}{change input e.g., ("=")}

\item{x}{the variable to change (e.g., year)}

\item{sql_code}{the sql query code...}

\item{flag}{a flag in the sql code to place the precode and x in the appropriate location}
}
\description{
utility function to filter sql files
}
\examples{
\dontrun{
.d = sql_filter(sql_precode = "<=", 2011, sql_code = .d, flag = "-- insert year")
}
}
