% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_read}
\alias{sql_read}
\title{read sql file}
\usage{
sql_read(x)
}
\arguments{
\item{x}{the sql code to read, pulled from the top directory}
}
\value{

}
\description{
read sql file
}
\examples{
\dontrun{
.d = sql_read("fsh_catch.sql")
}
}
