% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pcod_catch_1977_1990}
\alias{goa_pcod_catch_1977_1990}
\title{GOA Pacific cod historical catch}
\format{
A data frame with 280 observations and 4 variables:
\describe{
\item{tons}{catch in tons}
\item{gear}{fishery gear type}
\item{year}{1977-1990}
\item{season}{quarterly fishery season}
}
}
\usage{
goa_pcod_catch_1977_1990
}
\description{
A dataset containing historical catch data by year for 1977-1990
}
\keyword{datasets}
