% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_bts_biomass}
\alias{q_bts_biomass}
\title{query bottom trawl survey biomass}
\usage{
q_bts_biomass(
  year,
  area,
  species,
  by = "total",
  db,
  print_sql = FALSE,
  save = TRUE
)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{area}{options are bs, bsslope, nbs, ai, goa, old_bs - can only call a single area}

\item{species}{5 digit afsc species code(s) e.g., 79210 or c(79210, 10110)}

\item{by}{"depth", "stratum", "area", "total", "inpfc", "inpfc_depth" - only available for goa/ai (default: "total")}

\item{db}{the database to query (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\description{
probably need to beef up the documentation on the "by" switch
}
\details{
bs = bering sea + northwest 1987-present (includes nw stations) - recommended
bsslope = bering sea slope
nbs = northern bering sea
ai = aleutian islands
goa = gulf of alaska
old_bs = bering sea standard 1982-present (minus ~20 stations in nw) - not recommended
}
