% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_catch.R
\name{q_catch}
\alias{q_catch}
\title{query fishery catch data}
\usage{
q_catch(year, species, area, db, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{species group code e.g., "DUSK"}

\item{area}{sample area "GOA", "AI" or "BS" (or combos)}

\item{db}{data server to connect to}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\description{
query fishery catch data
}
\examples{
\dontrun{
q_catch(year = 2022, species = 'NORK', area = 'GOA', db = akfin, save = TRUE)
}
}
