% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_catch}
\alias{q_catch}
\title{query fishery catch data}
\usage{
q_catch(year, species, area, db, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{species group code e.g., "DUSK"}

\item{area}{sample area "GOA", "AI" or "BS" (or combos)}

\item{db}{data server to connect to}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\value{
a fsh_catch_data.csv file to the data/raw folder
}
\description{
Note that this function will produce results that are confidential. They can be hidden from git/GitHub by adding fsh_catch_data.csv to your .gitignore file
}
\examples{
\dontrun{
# query northern rockfish in the goa and save the results in default location
q_catch(year = 2022, species = 'NORK', area = 'GOA', db = akfin, save = TRUE)
}
}
