% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_catch.R
\name{q_catch}
\alias{q_catch}
\title{query fishery catch data}
\usage{
q_catch(year, species, area, db, group_code = TRUE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{species group code e.g., "DUSK". If group_code is set to
FALSE, input agency species code e.g., "145" = yelloweye rockfish}

\item{area}{FMP sample area "GOA" or "BSAI" (or both c("GOA", "BSAI"))}

\item{db}{data server to connect to}

\item{group_code}{defaults to TRUE}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\description{
query fishery catch data
}
\examples{
\dontrun{
q_catch(year = 2022, species = 'NORK', area = 'GOA', db = akfin, save = TRUE)
}
}
