% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{q_discards}
\alias{q_discards}
\title{query fishery discard data from AKFIN server}
\usage{
q_discards(year, species, area, db, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{species}{species group code e.g., "DUSK" or numeric agency values e.g. c("131", "132") - must be either all 4 digit or 3 digit codes}

\item{area}{fmp_area (GOA, BSAI) or fmp_subarea (BS, AI, WG, CG, WY, EY, SE) - also available (SEI, PWSI), can use all fmp_areas or all fmp_subareas, but don't mix the two}

\item{db}{data server to connect to (akfin)}

\item{save}{saves a file to the data/output folder, otherwise sends output to global enviro (default: TRUE)}
}
\value{
saves discard data as data/raw/fish_discard_data.csv or outputs to the global environment
}
\description{
query fishery discard data from AKFIN server
}
\examples{
\dontrun{
db <- afscdata::connect()
q_discards(year=2026, species="NORK", area="goa", db=db)
afscdata::disconnect(db)
}
 
}
