% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_fish_obs}
\alias{q_fish_obs}
\title{fishery observer data query}
\usage{
q_fish_obs(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{numeric agency values e.g. c("131", "132") - must be 3 digit codes (norpac species codes)}

\item{area}{fmp_area (GOA, BSAI) or fmp_subarea (BS, AI, WG, CG, WY, EY, SE) - also available (SEI, PWSI), can use all fmp_areas or all fmp_subareas, but don't mix the two}

\item{db}{data server to connect to (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data - save must be false}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: TRUE)}
}
\value{
saves observer data as data/raw/fsh_obs_data.csv or outputs to the global environment, save also saves a copy of the SQL code used for the query and stores it in the data/sql folder.
}
\description{
currently setup for GOA rockfish as unsure how others are estimating data through the end of the year
}
\examples{
\dontrun{
db <- afscdata::connect()
q_fish_obs(year=2022, species="NORK", area="goa", db=db)
}

}
