% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{q_for_catch}
\alias{q_for_catch}
\title{query foreign catch data from the AKFIN server}
\usage{
q_for_catch(year, species, area, db, print_sql = FALSE, save = TRUE)
}
\arguments{
\item{year}{max year to retrieve data from}

\item{species}{common name (e.g., "sablefish" or "all flounders") - can call multiple species}

\item{area}{numeric area digit code}

\item{db}{data server to connect to (akfin)}

\item{print_sql}{outputs the sql query instead of calling the data (default: false)}

\item{save}{saves a file to the data/raw folder, otherwise sends output to global enviro (default: true)}
}
\description{
something about the species names and area codes?
pulls data from akfin.pre1991.foreign_blend
}
